requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pandas
joblib
scikit-learn
pdfplumber
python-docx
psycopg2-binary
